Corrected logistic regression function
Hi guys, I corrected the logistic regression function. It was missing a column of ones in the matrix to calculate for the intercept. I compared it to the glm function and the resulting coefficients are very similar. Now we have to work in all other outputs the professors wants us to show.